kab: install dependent packages early and abort if failed
Install git before dealing with a git repo.

Install the dependent packages for building a kernel before make
localmodconfig.

When dependent packages fail to be installed, abort.

Note gcc-plugin-devel is not needed because latest
kernel-ark and RHEL disable CONFIG_GCC_PLUGINS. One benefit of dropping
gcc-plugin-devel is it could avoid install dependent packages for
building kernel because 1) gcc-plugin-devel in the CRB repo which is not
enabled sometimes 2) if there is one package missing, dnf would ruse to
install the remaining packages all together unless --skip-broken is
specified.

Signed-off-by: Coiby Xu <coxu@redhat.com>